Fix an integer overflow error for large files
Large files cause `"You tried to feed more data than fits in the file."`
error from Rubygems tar file writer.  The solution is to pass bytesize
instead of length to add_file_simple.

Note: I didn't observe this on a GNU/Linux machine but it was seen on a
Mac with a 10KiB resulting tgz file.  I'm guessing there is a difference
in byte size between int, long, and long long on the different systems.